services/backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==6.1.0
